refactor: Extract RecallInfoBox from Cars component
